+++ Attiny85/src/Storage.h
@@ -7,7 +7,7 @@
 #include "Setup.h"
 
 
-uint8_t crc_8(unsigned char *input_str, uint8_t num_bytes);
+uint8_t crc_8(unsigned char *input_str, size_t num_bytes);
 
 template<class T>
 class EEPROMStorage
